Migrate from .NET Framework 4.7.2 to .NET 10 (self-contained)
Convert legacy csproj files to SDK-style targeting net10.0-windows.
Add NuGet packages for APIs that moved out of base runtime. Fix
breaking changes: WindowsImpersonationContext removal (replaced with
P/Invoke), MenuItem/ContextMenu removal, MethodInvoker ambiguity,
File/Directory ACL API changes, and SendMessage nint overload ambiguity.

+++ Gw2Launcher/Client/Launcher_Autologin.cs
@@ -948,7 +948,7 @@ namespace Gw2Launcher.Client
 
                 Action wait = delegate
                 {
-                    NativeMethods.SendMessage(handle, 0, 0, 0);
+                    NativeMethods.SendMessage(handle, (uint)0, (uint)0, (uint)0);
                 };
 
                 //disabling to prevent interference from clicking (changes focus) - it'll still process keyboard input
